refactor: Use CommonJS with types in JSDoc
