+++ .pre-commit-config.yaml
@@ -4,9 +4,9 @@
 repos:
     - repo: local
       hooks:
-          - id: compile-typescript
-            name: Compile TypeScript
+          - id: check-typescript
+            name: Check TypeScript
             language: node
-            entry: npm run build
+            entry: npm exec tsc
             always_run: true
             pass_filenames: false
