+++ index.cjs
@@ -1,13 +1,14 @@
-import * as fs from "fs"
-import * as path from "path"
-import * as lockfile from "@yarnpkg/lockfile"
-import * as YAML from "yaml"
-import * as semverSort from "semver-sort"
+const fs = require("node:fs")
+const path = require("node:path")
+const lockfile = require("@yarnpkg/lockfile")
+const YAML = require("yaml")
+const semverSort = require("semver-sort")
 
-interface Dependency {
-    name: string
-    version: string | null
-}
+/**
+ * @typedef Dependency
+ * @property {string} name
+ * @property {string | null} version
+*/
 
 if (process.argv.length < 3) {
     throw new Error("Lock file as argument is required")
@@ -16,35 +17,60 @@ if (process.argv.length < 3) {
 const PRE_COMMIT_YAML = ".pre-commit-config.yaml"
 const LOCK_FILE = process.argv[2]
 
-const readFile = (path: string): string => fs.readFileSync(path, "utf8")
+/**
+ * @function
+ * @param {string} path
+ * @returns {string}
+ */
+const readFile = (path) => fs.readFileSync(path, "utf8")
 
-const parseVersion = (dependency: string): Dependency => {
+/**
+ * @param {string} dependency
+ * @returns {Dependency}
+ */
+function parseVersion (dependency) {
     const {name, version } = /^(?<name>@?[a-z-/0-9]+)(@(?<version>\d+\.\d+\.\d+))?$/.exec(dependency)?.groups || {name: "", version: null};
     return {
         name,
         version: version || null,
     }
-    // console.log({name, version});
-    const parts = dependency.split("@")
-    const lastIndex = parts.length - 1
-    return {
-        name: parts.slice(0, lastIndex).join("@"),
-        version: parts[lastIndex],
-    }
 }
 
-function parseYarnLockFile(file: string): Record<string, string[]> {
+/**
+ * @param {string} file
+ * @returns {Record<string, string[]>}
+ */
+function parseYarnLockFile(file) {
     const dependencies = lockfile.parse(readFile(file))
     return Object.keys(dependencies.object).reduce((mapping, dependency) => {
         const {name} = parseVersion(dependency)
         const installedVersion = dependencies.object[dependency].version
-        if (!mapping[name]) mapping[name] = []
-        mapping[name].push(installedVersion)
+        addElement(mapping, name, installedVersion)
         return mapping
-    }, {} as Record<string, string[]>)
+    }, {} /** @type {Record<string, string[]>} */)
 }
 
-function parseYarn3LockFile(file: string): Record<string, string[]> {
+/**
+ * Helper function for pushing another value to a record of lists.
+ * This helps with making TypeScript / JSDoc happy
+ *
+ * @template T
+ * @param {Record<string, T[]>} mapping
+ * @param {string} key
+ * @param {T} value
+ *
+ * @returns {void}
+ */
+function addElement(mapping, key, value) {
+    if (!(key in mapping)) mapping[key] = []
+    mapping[key].push(value)
+}
+
+/**
+ * @param {string} file
+ * @returns {Record<string, string[]>}
+ */
+function parseYarn3LockFile(file) {
     const dependencies = YAML.parse(readFile(file))
     if (![6, 8].includes(dependencies.__metadata.version)) {
         throw new Error("Unsupported format")
@@ -55,17 +81,27 @@ function parseYarn3LockFile(file: string): Record<string, string[]> {
         const parts =  dependency.split("@")
         const name = !parts[0] ? "@" + parts[1] : parts[0]
         const installedVersion = dependencies[dependency].version
-        if (!mapping[name]) mapping[name] = []
-        mapping[name].push(installedVersion)
+        addElement(mapping, name, installedVersion)
         return mapping
-    }, {} as Record<string, string[]>)
+    },
+        /** @type {Record<string, string[]>} */
+        {}
+    )
 }
 
-function parsePackageLockFile(file: string): Record<string, string[]> {
+/**
+ * @param {string} file
+ * @returns {Record<string, string[]>}
+ */
+function parsePackageLockFile(file) {
     const dependencies = JSON.parse(readFile(file))
     if (![2, 3].includes(dependencies.lockfileVersion)) throw new Error(`Unsupported version of package-lock.json (${dependencies.lockfileVersion})`)
 
-    function parseName(dependency: string): string {
+    /**
+     * @param {string} dependency
+     * @returns {string}
+     */
+    function parseName(dependency) {
         let [name, namespace, ...prefix] = dependency.split("/").reverse()
         if (namespace.startsWith("@")) {
             name = namespace + "/" + name
@@ -75,13 +111,18 @@ function parsePackageLockFile(file: string): Record<string, string[]> {
 
     return Object.keys(dependencies.packages).filter(name => !!name).reduce((mapping, dependency) => {
         const name = parseName(dependency)
-        if (!(name in mapping)) mapping[name] = []
-        mapping[name].push(dependencies.packages[dependency].version)
+        addElement(mapping, name, dependencies.packages[dependency].version)
         return mapping
-    }, ({} as Record<string, string[]>))
+    },
+        /** @type {Record<string, string[]>} */
+        {}
+    )
 }
 
-function getDependencies(): Record<string, string[]> {
+/**
+ * @returns {Record<string, string[]>}
+ */
+function getDependencies(){
     const get = () => {
         switch (path.basename(LOCK_FILE)) {
             case "yarn.lock":
@@ -105,29 +146,37 @@ function getDependencies(): Record<string, string[]> {
 
 const dependencies = getDependencies()
 
-function updateDependencies(): void {
+/**
+ * @returns {void}
+ */
+function updateDependencies() {
+
+    /**
+     * @typedef Hook
+     * @property {string} id
+     * @property {string[]} additional_dependencies
+     *
+     * @typedef Repo
+     * @property {string} repo
+     * @property {Hook[]} hooks
+     *
+     * @typedef PreCommit
+     * @property {Repo[]} repos
+    */
+
+    /** @type {PreCommit} */
     const preCommit = YAML.parse(readFile(PRE_COMMIT_YAML))
 
-    interface Hook {
-        id: string
-        additional_dependencies: string[]
-    }
-
-    interface Repo {
-        repo: string
-        hooks: Hook[]
-    }
-
-    interface PreCommit {
-        repos: Repo[]
-    }
-
-    const eslintHook = (preCommit as PreCommit).repos
+    const eslintHook = preCommit.repos
         .find((obj) => obj.repo === "https://github.com/pre-commit/mirrors-eslint")
         ?.hooks.find((hook) => hook.id === "eslint")
     if (eslintHook?.additional_dependencies) {
         const mapping = eslintHook.additional_dependencies.reduce(
-            (mapping, dependency) => {
+            (
+            /** @type {Record<string, string>} */
+                mapping,
+             dependency
+            ) => {
                 const {name} = parseVersion(dependency)
                 const installedVersions = dependencies[name]
                 if (!name) {
@@ -138,7 +187,8 @@ function updateDependencies(): void {
                     : dependency
                 return mapping
             },
-            {} as Record<string, string>
+            /** @type {Record<string, string>} */
+            {}
         )
         let content = readFile(PRE_COMMIT_YAML)
         Object.keys(mapping).forEach((previousVersion) => {
