+++ tsconfig.json
@@ -1,10 +1,14 @@
 {
+    "files": ["index.cjs"],
     "compilerOptions": {
         "target": "ESNext",
         "module": "CommonJS",
         "lib": ["ESNext", "DOM"],
         "types": ["node", "semver-sort"],
         "moduleResolution": "Node",
+        "allowJs": true,
+        "checkJs": true,
+        "noEmit": true,
         "isolatedModules": false,
         "strict": true,
         "esModuleInterop": true,
